Tentei centralizar o mapa, mudei a localização default para São José do Rio Preto e mandei o api da google pra um componente
